modification

+++ Reconcillations/Pages/ImpBank/Index.cshtml.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Reconcillations.Entity;
@@ -20,8 +21,10 @@ namespace Reconcillations.Pages.ImpBank
 
         [BindProperty]
         public List<TransactionList> _translist { get; set; }
+
         [TempData]
         public string Message { get; set; }
+
         public void OnGet()
         {
             _translist = _transactionRepository.GeTransactionLists();
@@ -31,21 +34,28 @@ namespace Reconcillations.Pages.ImpBank
         {
             ResponseInfo responses = new ResponseInfo();
 
-            if (id > 0)
+            string usertype = HttpContext.Session.GetString("UserType");
+
+            if (usertype != "I")
             {
-                responses = _transactionRepository.Deletereconciletransactionlist(id);
-
-                if (responses.StatusCode == "00")
+                if (id > 0)
                 {
-                    Message = responses.StatusMessage;
+                    responses = _transactionRepository.Deletereconciletransactionlist(id);
 
-                    return RedirectToPage("/Reconcile/Index");
-                }
-                else
-                {
-                    Message = responses.StatusMessage;
+                    if (responses.StatusCode == "00")
+                    {
+                        Message = responses.StatusMessage;
+
+                        return RedirectToPage("/Reconcile/Index");
+                    }
+                    else
+                    {
+                        Message = responses.StatusMessage;
+                    }
                 }
             }
+
+
 
             return Page();
